app/requirements.txt
llama-index-core==0.10.39
llama-index-embeddings-openai==0.1.5
llama-index-readers-file
llama-index-vector-stores-chroma==0.1.4
langchain==0.1.14
langchain-community==0.0.30
chromadb[server]==0.4.24
streamlit==1.33.0
gradio==4.15.0
watchdog==3.0.0
gitpython==3.1.43
pydantic-settings>=2.0.0
python-dotenv==1.0.1
uvicorn==0.28.0
